# Week 4_Web API_HandsOn/5.Web API_handsOn/testApiForJWT/Models/Employee.cs
using System;
using System.Collections.Generic;

namespace testApiForJWT.Models
{
    public class Employee
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Salary { get; set; }
        public bool Permanent { get; set; }
        public Department Department { get; set; }
        public List<Skill> Skills { get; set; }
        public DateTime DateOfBirth { get; set; }
    }
}